change encapsulation of encryption logic
EncryptedCapitalTransferInfo contained encrypted data and the encryption
logic - not as such a bad combination, but made somewhat confusing by
the very mutable state.

The new approach seems cleaner - with the drawback that the rest
endpoints are slightly less explicit (they use the general EncryptedData
instead of EncryptedCapitalTransferInfo - but even then the caller must
know which decryption logic to use, as no type info is transferred via
rest)
