+++ encrypted-data-model/src/main/java/ch/prevo/open/encrypted/model/EncryptedData.java
@@ -1,0 +1,56 @@
+package ch.prevo.open.encrypted.model;
+
+import java.io.Serializable;
+import java.util.Objects;
+
+public class EncryptedData implements Serializable {
+
+    private String encryptedDataBase64;
+    private String encryptedSymmetricKeyBase64;
+
+    public EncryptedData() {
+    }
+
+    public EncryptedData(String encryptedDataBase64, String encryptedSymmetricKeyBase64) {
+        this.encryptedDataBase64 = encryptedDataBase64;
+        this.encryptedSymmetricKeyBase64 = encryptedSymmetricKeyBase64;
+    }
+
+    public String getEncryptedDataBase64() {
+        return encryptedDataBase64;
+    }
+
+    public String getEncryptedSymmetricKeyBase64() {
+        return encryptedSymmetricKeyBase64;
+    }
+
+    public void setEncryptedDataBase64(String encryptedDataBase64) {
+        this.encryptedDataBase64 = encryptedDataBase64;
+    }
+
+    public void setEncryptedSymmetricKeyBase64(String encryptedSymmetricKeyBase64) {
+        this.encryptedSymmetricKeyBase64 = encryptedSymmetricKeyBase64;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        EncryptedData that = (EncryptedData) o;
+        return Objects.equals(encryptedDataBase64, that.encryptedDataBase64) &&
+                Objects.equals(encryptedSymmetricKeyBase64, that.encryptedSymmetricKeyBase64);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(encryptedDataBase64, encryptedSymmetricKeyBase64);
+    }
+
+    @Override
+    public String toString() {
+        return "EncryptedData{" +
+                "encryptedDataBase64='" + encryptedDataBase64 + '\'' +
+                ", encryptedSymmetricKeyBase64='" + encryptedSymmetricKeyBase64 + '\'' +
+                '}';
+    }
+}
