+++ hub/src/test/java/ch/prevo/open/hub/nodes/NodeServiceTest.java
@@ -1,6 +1,6 @@
 package ch.prevo.open.hub.nodes;
 
-import ch.prevo.open.encrypted.model.EncryptedCapitalTransferInfo;
+import ch.prevo.open.encrypted.model.EncryptedData;
 import ch.prevo.open.encrypted.model.InsurantInformation;
 import ch.prevo.open.encrypted.model.MatchForTermination;
 import ch.prevo.open.hub.match.Match;
@@ -121,7 +121,7 @@ public class NodeServiceTest {
     public void notifyMatch() {
         when(nodeRegistry.getCurrentNodes()).thenReturn(asList(node1_new, node2_old));
 
-        EncryptedCapitalTransferInfo transferInformation = new EncryptedCapitalTransferInfo();
+        EncryptedData transferInformation = new EncryptedData();
         when(nodeCaller.postCommencementNotification(eq(node2_old.getCommencementMatchNotifyUrl()), any()))
                 .thenReturn(transferInformation);
