+++ node/src/main/java/ch/prevo/open/node/api/MatchNotificationController.java
@@ -1,7 +1,6 @@
 package ch.prevo.open.node.api;
 
-import ch.prevo.open.encrypted.model.CapitalTransferInformation;
-import ch.prevo.open.encrypted.model.EncryptedCapitalTransferInfo;
+import ch.prevo.open.encrypted.model.EncryptedData;
 import ch.prevo.open.encrypted.model.MatchForCommencement;
 import ch.prevo.open.encrypted.model.MatchForTermination;
 import ch.prevo.open.node.data.provider.error.NotificationException;
@@ -31,12 +30,12 @@ public class MatchNotificationController {
     }
 
     @RequestMapping(value = "/commencement-match-notification", method = RequestMethod.POST)
-    public ResponseEntity<EncryptedCapitalTransferInfo> receiveCommencementMatchNotification(@RequestBody MatchForCommencement matchNotification) {
+    public ResponseEntity<EncryptedData> receiveCommencementMatchNotification(@RequestBody MatchForCommencement matchNotification) {
         LOGGER.debug("Receive commencement match notification for OASI {}, switching to new retirement fund: {}",
                 matchNotification.getEncryptedOasiNumber(), matchNotification.getRetirementFundUid());
 
         try {
-            final Optional<EncryptedCapitalTransferInfo> transferInformation = notificationService
+            final Optional<EncryptedData> transferInformation = notificationService
                     .handleTerminationMatch(matchNotification);
 
             return transferInformation.isPresent() ?
