+++ node/src/main/java/ch/prevo/open/node/services/MatchNotificationService.java
@@ -1,23 +1,19 @@
 package ch.prevo.open.node.services;
 
-import ch.prevo.open.data.api.FullCommencementNotification;
-import ch.prevo.open.data.api.FullTerminationNotification;
-import ch.prevo.open.data.api.EmploymentTermination;
-import ch.prevo.open.data.api.EmploymentInfo;
-import ch.prevo.open.data.api.EmploymentCommencement;
+import ch.prevo.open.data.api.*;
 import ch.prevo.open.encrypted.model.CapitalTransferInformation;
-import ch.prevo.open.encrypted.model.EncryptedCapitalTransferInfo;
-import ch.prevo.open.encrypted.model.MatchForTermination;
+import ch.prevo.open.encrypted.model.EncryptedData;
 import ch.prevo.open.encrypted.model.MatchForCommencement;
+import ch.prevo.open.encrypted.model.MatchForTermination;
+import ch.prevo.open.encrypted.services.CapitalTransferInfoEncrypter;
 import ch.prevo.open.encrypted.services.Cryptography;
 import ch.prevo.open.node.config.AdapterServiceConfiguration;
 import ch.prevo.open.node.config.NodeConfigurationService;
-import ch.prevo.open.node.data.provider.EmploymentTerminationProvider;
 import ch.prevo.open.node.data.provider.EmploymentCommencementProvider;
+import ch.prevo.open.node.data.provider.EmploymentTerminationProvider;
 import ch.prevo.open.node.data.provider.MatchNotificationListener;
 import ch.prevo.open.node.data.provider.ProviderFactory;
 import ch.prevo.open.node.data.provider.error.NotificationException;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.serviceloader.ServiceListFactoryBean;
@@ -61,16 +57,16 @@ public class MatchNotificationService {
         fullNotification.setCommencementDate(notification.getCommencementDate());
         fullNotification.setEmploymentTermination(employmentTermination.get());
 
-        EncryptedCapitalTransferInfo encryptedCapitalTransferInfo = notification.getTransferInformation();
+        EncryptedData encryptedCapitalTransferInfo = notification.getTransferInformation();
         if (encryptedCapitalTransferInfo != null) {
             String retirementFundUid = employmentTermination.get().getEmploymentInfo().getRetirementFundUid();
-            fullNotification.setTransferInformation(encryptedCapitalTransferInfo.decryptData(nodeConfigService.getPrivateKey(retirementFundUid)));
+            fullNotification.setTransferInformation(new CapitalTransferInfoEncrypter().decrypt(encryptedCapitalTransferInfo, nodeConfigService.getPrivateKey(retirementFundUid)));
         }
 
         listener.handleCommencementMatch(fullNotification);
     }
 
-    public Optional<EncryptedCapitalTransferInfo> handleTerminationMatch(MatchForCommencement notification)
+    public Optional<EncryptedData> handleTerminationMatch(MatchForCommencement notification)
             throws NotificationException {
         final Optional<EmploymentCommencement> employmentCommencement = employmentCommencementProvider.getEmploymentCommencements().stream()
                 .filter(currentEmploymentCommencement -> isSameAsNotification(currentEmploymentCommencement, notification))
@@ -92,7 +88,8 @@ public class MatchNotificationService {
         if (info == null) {
             return Optional.empty();
         }
-        return Optional.of(new EncryptedCapitalTransferInfo(info, nodeConfigService.getPublicKey(notification.getPreviousRetirementFundUid())));
+
+        return Optional.of(new CapitalTransferInfoEncrypter().encrypt(info, nodeConfigService.getPublicKey(notification.getPreviousRetirementFundUid())));
     }
 
     private boolean isSameAsNotification(EmploymentCommencement employmentCommencement, MatchForCommencement notification) {
